split tests into packages to avoid timeout issue
